<refactor>(backoffice): move Backoffice to pages folder
