+++ src/routes/index.jsx
@@ -4,8 +4,8 @@ import { Routes, Route } from 'react-router-dom'
 import ClientLayout from '../Components/ClientLayout/ClientLayout'
 import Login from '../pages/Login/Login'
 import Home from '../pages/Home'
+import BackofficeContainer from '../pages/Backoffice/BackofficeContainer'
 import Transactions from '../Components/Transactions/Transactions'
-import BackofficeContainer from '../Components/Backoffice/BackofficeContainer'
 
 const AppRoutes = () => (
   <Routes>
